backend/requirements.txt
fastapi
pymongo
python-dotenv
uvicorn
requests
python-multipart
email-validator
passlib[bcrypt]
httpx # Added for making async HTTP requests
